KeystrokeHelper introduced to avoid build server error "No X11 DISPLAY variable was set, but this program performed an operation which requires it."
